+++ examples/resources/catalystcenter_template/resource.tf
@@ -1,7 +1,8 @@
 resource "catalystcenter_template" "example" {
-  project_id  = "12345678-1234-1234-1234-123456789012"
-  name        = "Template1"
-  description = "My description"
+  project_id   = "12345678-1234-1234-1234-123456789012"
+  name         = "Template1"
+  project_name = "ProjectName"
+  description  = "My description"
   device_types = [
     {
       product_family = "Switches and Hubs"
@@ -29,4 +30,5 @@ resource "catalystcenter_template" "example" {
       selection_values = { host1 = "host1" }
     }
   ]
+  composite = false
 }