+++ internal/provider/data_source_catalystcenter_template.go
@@ -75,6 +75,10 @@ func (d *TemplateDataSource) Schema(ctx context.Context, req datasource.SchemaRe
 				Optional:            true,
 				Computed:            true,
 			},
+			"project_name": schema.StringAttribute{
+				MarkdownDescription: "ProjectName",
+				Computed:            true,
+			},
 			"description": schema.StringAttribute{
 				MarkdownDescription: "Description",
 				Computed:            true,
@@ -192,6 +196,34 @@ func (d *TemplateDataSource) Schema(ctx context.Context, req datasource.SchemaRe
 						"selection_values": schema.MapAttribute{
 							MarkdownDescription: "Selection values",
 							ElementType:         types.StringType,
+							Computed:            true,
+						},
+					},
+				},
+			},
+			"composite": schema.BoolAttribute{
+				MarkdownDescription: "Is it composite template",
+				Computed:            true,
+			},
+			"containing_templates": schema.ListNestedAttribute{
+				MarkdownDescription: "Containing templates for composite template",
+				Computed:            true,
+				NestedObject: schema.NestedAttributeObject{
+					Attributes: map[string]schema.Attribute{
+						"name": schema.StringAttribute{
+							MarkdownDescription: "Name of the template",
+							Computed:            true,
+						},
+						"id": schema.StringAttribute{
+							MarkdownDescription: "ID of the template",
+							Computed:            true,
+						},
+						"project_name": schema.StringAttribute{
+							MarkdownDescription: "Project name",
+							Computed:            true,
+						},
+						"language": schema.StringAttribute{
+							MarkdownDescription: "Language of the template",
 							Computed:            true,
 						},
 					},