+++ internal/provider/resource_catalystcenter_template.go
@@ -82,6 +82,10 @@ func (r *TemplateResource) Schema(ctx context.Context, req resource.SchemaReques
 				MarkdownDescription: helpers.NewAttributeDescription("Name of the template").String,
 				Required:            true,
 			},
+			"project_name": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("ProjectName").String,
+				Optional:            true,
+			},
 			"description": schema.StringAttribute{
 				MarkdownDescription: helpers.NewAttributeDescription("Description").String,
 				Optional:            true,
@@ -209,6 +213,37 @@ func (r *TemplateResource) Schema(ctx context.Context, req resource.SchemaReques
 							MarkdownDescription: helpers.NewAttributeDescription("Selection values").String,
 							ElementType:         types.StringType,
 							Optional:            true,
+						},
+					},
+				},
+			},
+			"composite": schema.BoolAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Is it composite template").String,
+				Optional:            true,
+			},
+			"containing_templates": schema.ListNestedAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Containing templates for composite template").String,
+				Optional:            true,
+				NestedObject: schema.NestedAttributeObject{
+					Attributes: map[string]schema.Attribute{
+						"name": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Name of the template").String,
+							Required:            true,
+						},
+						"id": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("ID of the template").String,
+							Optional:            true,
+						},
+						"project_name": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Project name").String,
+							Required:            true,
+						},
+						"language": schema.StringAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Language of the template").AddStringEnumDescription("JINJA", "VELOCITY").String,
+							Required:            true,
+							Validators: []validator.String{
+								stringvalidator.OneOf("JINJA", "VELOCITY"),
+							},
 						},
 					},
 				},